chore: 회원가입 로직 위치 수정
