+++ src/main/java/com/dearnewyear/dny/user/controller/AuthController.java
@@ -3,12 +3,16 @@ package com.dearnewyear.dny.user.controller;
 import com.dearnewyear.dny.common.dto.response.ApiResponse;
 import com.dearnewyear.dny.common.dto.response.ErrorResponse;
 import com.dearnewyear.dny.common.error.exception.CustomException;
+import com.dearnewyear.dny.user.dto.request.SignupRequest;
 import com.dearnewyear.dny.user.service.KakaoOAuth2Service;
+import com.dearnewyear.dny.user.service.UserService;
 import java.io.IOException;
 import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,10 +20,23 @@ import org.springframework.web.servlet.view.RedirectView;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/oauth")
-public class OauthController {
+@RequestMapping("/auth")
+public class AuthController {
 
     private final KakaoOAuth2Service kakaoOAuth2Service;
+    private final UserService userService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<ApiResponse> signup(@RequestBody SignupRequest request, HttpServletResponse response) {
+        try {
+            userService.signupAndLoginUser(request, response);
+            ApiResponse res = new ApiResponse(200, "회원가입 성공", null);
+            return ResponseEntity.ok(res);
+        } catch (CustomException e) {
+            ErrorResponse res = ErrorResponse.of(e.getErrorCode());
+            return ResponseEntity.status(e.getErrorCode().getStatus()).body(res);
+        }
+    }
 
     @GetMapping("/login/kakao")
     public RedirectView getAuthorizationCode() throws IOException {
