+++ src/main/java/com/dearnewyear/dny/user/service/UserService.java
@@ -20,14 +20,28 @@ public class UserService {
     private final JwtTokenProvider jwtTokenProvider;
     private final UserRepository userRepository;
 
+    public void signupAndLoginUser(SignupRequest request, HttpServletResponse response) {
+        if (userRepository.findByEmail(request.getEmail()).isPresent())
+            throw new CustomException(ErrorCode.USER_ALREADY_EXIST);
+
+        User user = new User(request);
+        userRepository.save(user);
+
+        String accessToken = jwtTokenProvider.createAccessToken(user);
+        String refreshToken = jwtTokenProvider.createRefreshToken(user);
+
+        response.setHeader("Authorization", "Bearer " + accessToken);
+        response.setHeader("DNY-Refresh", refreshToken);
+    }
+
     public String chkAndLoginUser(String email, HttpServletResponse response) {
         Optional<User> user = userRepository.findByEmail(email);
         if (user.isPresent()) {
             String accessToken = jwtTokenProvider.createAccessToken(user.get());
             String refreshToken = jwtTokenProvider.createRefreshToken(user.get());
 
-            response.setHeader("Authorization", accessToken);
-            response.setHeader("RefreshToken", refreshToken);
+            response.setHeader("Authorization", "Bearer " + accessToken);
+            response.setHeader("DNY-Refresh", refreshToken);
             return "success";
         } else {
             return email;
